+++ src/individual/Elephant-Delirium/internal-util.c
@@ -23,9 +23,8 @@
 #include "internal-util.h"
 
 int aead_check_tag
-    (unsigned char *plaintext, unsigned long long plaintext_len,
-     const unsigned char *tag1, const unsigned char *tag2,
-     unsigned size)
+    (unsigned char *plaintext, size_t plaintext_len,
+     const unsigned char *tag1, const unsigned char *tag2, size_t size)
 {
     /* Set "accum" to -1 if the tags match, or 0 if they don't match */
     int accum = 0;
